move all fixtures to conftest
